Reorganize extension methods to align with Firebolt documentation structure
- Create LambdaMethods.cs for lambda-based array functions (ArrayTransform, ArrayFilter, ArrayAnyMatch, etc.)
- Create DataSketchesMethods.cs for HyperLogLog functions (HllCountBuild, HllCountEstimate, HllCountMerge)
- Rename FireboltSpecificExtensions.cs to DateTimeMethods.cs for date/time functions
- Remove moved functions from AggregationMethods.cs and ArrayMethods.cs
- Create corresponding test files: LambdaTests.cs, DataSketchesTests.cs, DateTimeTests.cs
- Update test files to remove moved tests (ArrayTests.cs, AggregationTests.cs, FireboltPoCTests.cs)
- Update references in MappingSchema.cs to use DateTimeMethods
