+++ src/Similarweb.LinqToDb.Firebolt/MappingSchema.cs
@@ -30,16 +30,16 @@ internal class MappingSchema : global::LinqToDB.Mapping.MappingSchema
     {
         SetDataType(typeof(DateTime), DataType.DateTime);
 
-        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Year), (DateTime obj) => FireboltSpecificExtensions.DatePart(Sql.DateParts.Year, obj)!.Value);
-        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Month), (DateTime obj) => FireboltSpecificExtensions.DatePart(Sql.DateParts.Month, obj)!.Value);
-        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.DayOfYear), (DateTime obj) => FireboltSpecificExtensions.DatePart(Sql.DateParts.DayOfYear, obj)!.Value);
-        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Day), (DateTime obj) => FireboltSpecificExtensions.DatePart(Sql.DateParts.Day, obj)!.Value);
-        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.DayOfWeek), (DateTime obj) => FireboltSpecificExtensions.DatePart(Sql.DateParts.WeekDay, obj)!.Value - 1);
-        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Hour), (DateTime obj) => FireboltSpecificExtensions.DatePart(Sql.DateParts.Hour, obj)!.Value);
-        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Minute), (DateTime obj) => FireboltSpecificExtensions.DatePart(Sql.DateParts.Minute, obj)!.Value);
-        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Second), (DateTime obj) => FireboltSpecificExtensions.DatePart(Sql.DateParts.Second, obj)!.Value);
+        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Year), (DateTime obj) => DateTimeMethods.DatePart(Sql.DateParts.Year, obj)!.Value);
+        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Month), (DateTime obj) => DateTimeMethods.DatePart(Sql.DateParts.Month, obj)!.Value);
+        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.DayOfYear), (DateTime obj) => DateTimeMethods.DatePart(Sql.DateParts.DayOfYear, obj)!.Value);
+        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Day), (DateTime obj) => DateTimeMethods.DatePart(Sql.DateParts.Day, obj)!.Value);
+        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.DayOfWeek), (DateTime obj) => DateTimeMethods.DatePart(Sql.DateParts.WeekDay, obj)!.Value - 1);
+        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Hour), (DateTime obj) => DateTimeMethods.DatePart(Sql.DateParts.Hour, obj)!.Value);
+        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Minute), (DateTime obj) => DateTimeMethods.DatePart(Sql.DateParts.Minute, obj)!.Value);
+        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now.Second), (DateTime obj) => DateTimeMethods.DatePart(Sql.DateParts.Second, obj)!.Value);
 
-        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now), () => FireboltSpecificExtensions.CurrentDateTime);
+        Expressions.MapMember(DataProvider.V2Id, Expressions.M(() => DateTime.Now), () => DateTimeMethods.CurrentDateTime);
 
         // NB: These converters are applied only for logs. Queries are built in FireboltCommand class.
         //     Main idea is that we shouldn't be too concerned with possible issues here, see tests.
